requirements.txt
playwright==1.40.0
redis==5.0.1
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
psycopg2-binary==2.9.9